mediacapture-streams/MediaStream-MediaElement-preload-none-manual.https.html

<!DOCTYPE html>
<title>Test that the HTMLMediaElement preload 'none' attribute value is ignored for MediaStream used as srcObject and MediaStream object URLs used as src.</title>
<link rel="author" title="Matthew Wolenetz" href="mailto:wolenetz@chromium.org"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p class="instructions">When prompted, accept to share your audio and video streams.</p>
<p class="instructions">This test checks that the HTMLMediaElement preload 'none' attribute value is ignored for MediaStream used as srcObject and MediaStream object URLs used as src.</p>

<audio preload="none"></audio>
<video preload="none"></video>

<script>

async function testPreloadNone(t, mediaElement, setSourceStreamFunc) {
  // The optional deferred load steps (for preload none) for MediaStream resources should be skipped.
  mediaElement.addEventListener("suspend", t.unreached_func("'suspend' should not be fired."));
  mediaElement.addEventListener("error", t.step_func(() =>
    assert_unreached("'error' should not be fired, code=" + mediaElement.error.code)));
  const watcher = new EventWatcher(t, mediaElement, ["loadeddata"]);
  setSourceStreamFunc();
  assert_equals(mediaElement.networkState, mediaElement.NETWORK_NO_SOURCE); // Resource selection is active.
  await watcher.wait_for("loadeddata");
  assert_equals(mediaElement.networkState, mediaElement.NETWORK_LOADING);
}

promise_test(async t => {
  const aud = document.querySelector("audio");
  const stream = await navigator.mediaDevices.getUserMedia({audio: true});
  t.add_cleanup(() => stream.getAudioTracks()[0].stop());
  await testPreloadNone(t, aud, () => { aud.srcObject = stream; });
}, "Test that preload 'none' is ignored for MediaStream object URL used as srcObject for audio");

promise_test(async t => {
  const vid = document.querySelector("video");
  const stream = await navigator.mediaDevices.getUserMedia({video: true});
  t.add_cleanup(() => stream.getVideoTracks()[0].stop());
  await testPreloadNone(t, vid, () => { vid.srcObject = stream; });
}, "Test that preload 'none' is ignored for MediaStream used as srcObject for video");

</script>
